requirements.txt
asgiref==3.3.4
Django==3.2
django-appconf==1.0.4
django-imagekit==4.0.2
django-resized==0.3.11
django-tinymce==3.3.0
gunicorn==20.1.0
jsmin==2.2.2
pilkit==2.0
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
